requirements.txt
# Dependencias principales del RPA
# Instalación: pip install -r requirements.txt

# Manipulación y análisis de datos
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Visualización de datos
matplotlib==3.8.2
seaborn==0.13.0

# Comunicación con WhatsApp
twilio==8.10.3

# Gestión de variables de entorno
python-dotenv==1.0.0

# Utilidades adicionales
requests==2.31.0